Move methods to userspace
